move span field logic to a separate module
